requirements.txt
# Core ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Healthcare specific
pydicom>=2.4.0
SimpleITK>=2.2.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Visualization and reporting
plotly>=5.15.0
bokeh>=3.2.0
wordcloud>=1.9.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: For specific project types
# For RL projects:
# gymnasium>=0.29.0
# stable-baselines3>=2.0.0

# For explainable AI:
# lime>=0.2.0
# shap>=0.42.0

# For privacy/synthetic data:
# sdv>=1.2.0
# diffprivlib>=0.6.0 